Streamline dockerfile and compose file
The way the dockerfile was constructed, the database is always copied into the image instead of being mounted from a volume. Due to the relative sqlite path
the db is always referenced from /app/data instead of from /data.

This change removes the mounts to clarify this and to minimize complexity while testing deployment to Azure AppService.

For the same reason the port has been redefined to 5000, since it is the default AppService port. This may also be updated later.

Finally, the streamlit parameters have been changed to environment variables which are easier to change at deploy time.