Add Azure Container Registry login to docker build workflow
The username and password for `docker/login-action` are stored as repository level vars
and secrets respectively. They reference an access token on the Azure Container Registry.
The name of the access token is used as username.

The access token password can not be retrieved again after creation, neither from the
Azure Portal nor from the Github secret. If you need to change the password, you need
to regenerate the password for the token in the Azure Portal, and then update the secret
in the Github repository secret.